+++ .vitepress/tools/design.js
@@ -3,8 +3,14 @@ export default {
   defaultExpanded: false,
   items: [
     {
+      link:'https://asciitree.com/',
       title:'ascii-tree-generator',
-      desc:'ascii树生成器',
+      desc:'ascii树生成器，高度自定义',
+      linktxt:'ascii-tree-generator.com'
+    },
+    {
+      title:'ascii-tree-generator',
+      desc:'ascii树生成器，可拖动',
       link:'https://ascii-tree-generator.com/',
       linktxt:'ascii-tree-generator.com'
     },